Make interface for AttributeInstance and AttributeMap
